requirements.txt
Django==4.0.5

# Models
# -------------------------------------
psycopg2==2.9.6
django-extensions==3.2.0
django-uuid-upload-path==1.0.0
django-versatileimagefield==3.0
Pillow==9.4.0
django-simple-history==3.3.0
dj-database-url==2.1.0



# REST APIs
# -------------------------------------
coreapi==2.3.3
djangorestframework==3.13.1
django-rest-swagger==2.2.0
darkskylib==0.3.91


# Debugging
# -------------------------------------
django-debug-toolbar==4.1.0
ipdb==0.13.13


django-bulk-update-or-create==0.3.0

django-environ==0.10.0
python-decouple==3.8

gunicorn==21.2.0